+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Adding default expression argument to `lead` & `lag` window functions.
 * Adding `shlex_split`, `cmd` and `shell` moonblade functions.
 * Adding `aarch64-apple-darwin` and `aarch64-unknown-linux-gnu` to CI builds.
+* Adding `to_fixed` moonblade function.
 
 *Fixes*